Factor globe into separate controller and rendering processes.
The current setup has been a bit flaky, sometimes freezing on button-presses and
otherwise not being an amazing user experience, especially for a wee one. This
change separates the control of the light and LCD screen into one process, and
control of the mode and other "higher-level" concepts into another process.

Needs to be tested, but hopefully the basic skeleton is there.